+++ tf2jModel/converter.py
@@ -1,9 +1,9 @@
 import tensorflow as tf
 import pickle
-from jmodel import Converter_runtime
+from jModel import Converter_runtime
 
 class Converter:
-    version = "0.0.1"
+    version = "0.0.2"
 
     def __init__(self):
         self.hyperparameter_separator = "--"
@@ -24,6 +24,8 @@ class Converter:
             weight_index += 1
             if layer.__class__.__name__ == "Dense":
                 layer_name = self.layer_name_add(layer_name, "Dense")
+                # Check Activation
+                layer_name = self.layer_name_add(layer_name, f"activation='{layer.activation.__name__}'")
                 # Check Bias
                 if layer.use_bias == True:
                     layer_name = self.layer_name_add(layer_name, "use_bias=True")
@@ -31,8 +33,24 @@ class Converter:
                     weight_index += 1
                 else:
                     layer_name = self.layer_name_add(layer_name, "use_bias=False")
+            elif layer.__class__.__name__ == "LSTM":
+                layer_name = self.layer_name_add(layer_name, "LSTM")
+                # Add Recurrent Kernel Weight
+                layer_weight.append(weights[weight_index])
+                weight_index += 1
+                # Check Units
+                layer_name = self.layer_name_add(layer_name, f"units='{layer.units}'")
                 # Check Activation
                 layer_name = self.layer_name_add(layer_name, f"activation='{layer.activation.__name__}'")
+                # Check Bias
+                if layer.use_bias == True:
+                    layer_name = self.layer_name_add(layer_name, "use_bias=True")
+                    layer_weight.append(weights[weight_index])
+                    weight_index += 1
+                else:
+                    layer_name = self.layer_name_add(layer_name, "use_bias=False")
+                # Check Return Sequences
+                layer_name = self.layer_name_add(layer_name, f"return_sequences={layer.return_sequences}")
             else:
                 raise TypeError(f"Layer {layer.__class__.__name__} is not supported for jModel=={Converter.version}")
             self.layers.append(layer_name)