+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2024-03-24
+## v0.0.2 - 2024-03-24
+
+### Added
+- Added compatibility for **LSTM** layers with hyperparameters: **units**, **activation**, **use_bias**, **return_sequences**
+- Added Softmax Activation Function
+
+### Changed
+- Changed the layer hyperparameters sequence for tf2jModel to same as TensorFlow
+
+## v0.0.1 - 2024-03-24
 
 ### Added